work in progress on changing nan-boxing to pointer-tagging
This is for better future compatibility with the WebAssembly
garbage collection extension proposal.

In the GC proposal, you can't alias GC references with integers
so NaN-boxing works only for references to linear memory.

But there is a provision for an 'int31ref' type which is a tagged
pointer-compatible value which stores a 31-bit integer. This can
be implemented on a 32-bit arch by using the lowest bit as a tag,
since it will never be set on pointers (always 4-byte or 8-byte
aligned structures, so bottom 2 bits are free!)

Switching to a similar system for working within linear memory:
* values are a pointer-sized word (so 32-bit for wasm32)
* object references store the pointer directly
* integers in 31-bit range are stored shifted with a tag bit
* special values (undefined, true, etc) are refs to singleton objs
* doubles and integers outside 31-bit range are stored as boxed objs

This does mean that heavy math code could be ugly/slow. But that's
not an expected primary use case for this project, and it'd be
more interesting to work with the future native GC support.